+++ packages/go_router/lib/src/match.dart
@@ -74,7 +74,7 @@ class RouteMatch {
   /// An exception if there was an error during matching.
   final Exception? error;
 
-  /// Optional value key of type string, to hold a unique reference to a page.
+  /// Value key of type string, to hold a unique reference to a page.
   final ValueKey<String> pageKey;
 
   @override