+++ packages/go_router/lib/src/route_data.dart
@@ -14,7 +14,7 @@ import 'state.dart';
 /// Baseclass for supporting
 /// [Type-safe routing](https://pub.dev/documentation/go_router/latest/topics/Type-safe%20routes-topic.html).
 ///
-/// Subclasses must override one of [build], [buildPageWithState], or
+/// Subclasses must override one of [build], [buildPage], or
 /// [redirect].
 /// {@category Type-safe routes}
 abstract class GoRouteData {
@@ -25,53 +25,36 @@ abstract class GoRouteData {
 
   /// Creates the [Widget] for `this` route.
   ///
-  /// Subclasses must override one of [build], [buildPageWithState], or
+  /// Subclasses must override one of [build], [buildPage], or
   /// [redirect].
   ///
   /// Corresponds to [GoRoute.builder].
-  Widget build(BuildContext context) => throw UnimplementedError(
-        'One of `build` or `buildPageWithState` must be implemented.',
+  Widget build(BuildContext context, GoRouterState state) =>
+      throw UnimplementedError(
+        'One of `build` or `buildPage` must be implemented.',
       );
 
   /// A page builder for this route.
   ///
   /// Subclasses can override this function to provide a custom [Page].
   ///
-  /// Subclasses must override one of [build], [buildPageWithState] or
+  /// Subclasses must override one of [build], [buildPage] or
   /// [redirect].
   ///
   /// Corresponds to [GoRoute.pageBuilder].
   ///
   /// By default, returns a [Page] instance that is ignored, causing a default
   /// [Page] implementation to be used with the results of [build].
-  @Deprecated(
-    'This method has been deprecated in favor of buildPageWithState. '
-    'This feature was deprecated after v4.3.0.',
-  )
-  Page<void> buildPage(BuildContext context) => const NoOpPage();
-
-  /// A page builder for this route with [GoRouterState].
-  ///
-  /// Subclasses can override this function to provide a custom [Page].
-  ///
-  /// Subclasses must override one of [build], [buildPageWithState] or
-  /// [redirect].
-  ///
-  /// Corresponds to [GoRoute.pageBuilder].
-  ///
-  /// By default, returns a [Page] instance that is ignored, causing a default
-  /// [Page] implementation to be used with the results of [build].
-  Page<void> buildPageWithState(BuildContext context, GoRouterState state) =>
-      // ignore: deprecated_member_use_from_same_package
-      buildPage(context);
+  Page<void> buildPage(BuildContext context, GoRouterState state) =>
+      const NoOpPage();
 
   /// An optional redirect function for this route.
   ///
-  /// Subclasses must override one of [build], [buildPageWithState], or
+  /// Subclasses must override one of [build], [buildPage], or
   /// [redirect].
   ///
   /// Corresponds to [GoRoute.redirect].
-  FutureOr<String?> redirect() => null;
+  FutureOr<String?> redirect(BuildContext context, GoRouterState state) => null;
 
   /// A helper function used by generated code.
   ///
@@ -106,13 +89,13 @@ abstract class GoRouteData {
     }
 
     Widget builder(BuildContext context, GoRouterState state) =>
-        factoryImpl(state).build(context);
+        factoryImpl(state).build(context, state);
 
     Page<void> pageBuilder(BuildContext context, GoRouterState state) =>
-        factoryImpl(state).buildPageWithState(context, state);
+        factoryImpl(state).buildPage(context, state);
 
     FutureOr<String?> redirect(BuildContext context, GoRouterState state) =>
-        factoryImpl(state).redirect();
+        factoryImpl(state).redirect(context, state);
 
     return GoRoute(
       path: path,