+++ packages/pigeon/platform_tests/alternate_language_test_plugin/example/ios/RunnerTests/RunnerTests.m
@@ -11,8 +11,8 @@
 #endif
 
 @interface ACMessageSearchReply ()
-+ (ACMessageSearchReply *)fromMap:(NSDictionary *)dict;
-- (NSDictionary *)toMap;
++ (ACMessageSearchReply *)fromList:(NSArray *)list;
+- (NSArray *)toList;
 @end
 
 @interface RunnerTests : XCTestCase
@@ -24,24 +24,24 @@
 - (void)testToMapAndBack {
   ACMessageSearchReply *reply = [[ACMessageSearchReply alloc] init];
   reply.result = @"foobar";
-  NSDictionary *dict = [reply toMap];
-  ACMessageSearchReply *copy = [ACMessageSearchReply fromMap:dict];
+  NSArray *list = [reply toList];
+  ACMessageSearchReply *copy = [ACMessageSearchReply fromList:list];
   XCTAssertEqual(reply.result, copy.result);
 }
 
 - (void)testHandlesNull {
   ACMessageSearchReply *reply = [[ACMessageSearchReply alloc] init];
   reply.result = nil;
-  NSDictionary *dict = [reply toMap];
-  ACMessageSearchReply *copy = [ACMessageSearchReply fromMap:dict];
+  NSArray *list = [reply toList];
+  ACMessageSearchReply *copy = [ACMessageSearchReply fromList:list];
   XCTAssertNil(copy.result);
 }
 
 - (void)testHandlesNullFirst {
   ACMessageSearchReply *reply = [[ACMessageSearchReply alloc] init];
   reply.error = @"foobar";
-  NSDictionary *dict = [reply toMap];
-  ACMessageSearchReply *copy = [ACMessageSearchReply fromMap:dict];
+  NSArray *list = [reply toList];
+  ACMessageSearchReply *copy = [ACMessageSearchReply fromList:list];
   XCTAssertEqual(reply.error, copy.error);
 }